pyproject.toml
[project]
name = "bimi-preview"
version = "0.3.3"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1",
    "pillow>=12.0",
    "cairosvg>=2.9",
    "numpy>=2.0",
    "scipy>=1.14",
    "scikit-image>=0.24",
    "anthropic>=0.40",
    "openai>=1.50",
    "google-genai>=1.0",
    "python-dotenv>=1.0",
    "waitress>=3.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0"]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
